renaming the rooms-npc module to rooms_npc
